+++ packages/eslint-config-inpsyde/rules/nodejs-and-commonjs.js
@@ -7,6 +7,7 @@ module.exports = {
 		'callback-return': 0,
 		'global-require': 2,
 		'handle-callback-err': 0,
+		'no-buffer-constructor': 0,
 		'no-mixed-requires': 0,
 		'no-new-require': 2,
 		'no-path-concat': 2,